+++ core/types/row_consumption.go
@@ -7,8 +7,22 @@ import (
 	"github.com/scroll-tech/go-ethereum/rlp"
 )
 
+type SubCircuitRowUsage struct {
+	Name      string `json:"name"`
+	RowNumber uint64 `json:"row_number"`
+}
+
+type RowUsage struct {
+	IsOk            bool                 `json:"is_ok"`
+	RowNumber       uint64               `json:"row_number"`
+	RowUsageDetails []SubCircuitRowUsage `json:"row_usage_details"`
+}
+
 type RowConsumption struct {
 	Rows uint64
+
+	// tmp workaround. will only keep `Detail` and remove `Rows` later
+	Detail []SubCircuitRowUsage
 }
 
 func (rc *RowConsumption) EncodeRLP(w io.Writer) error {
